run animation smoother

+++ anim_lava.h
@@ -16,7 +16,7 @@ private:
   unsigned long last_hue_step;
   unsigned long last_bubble_step;
   uint8_t hue_pos;
-  uint8_t bubble_pos;
+  unsigned int bubble_pos;
   
   uint8_t sub_pixels;
   uint8_t dim_part;
